Added new files to repo
Biggest update here is Stringcrease. I wanted to print a string based on something the user entered as input, but didn't need the loop to contain IF or ELIF statements. The string prints whatever text the user enters letter by letter, concatenating one character for each line until it reaches the end of the string.

NaNoWriMo is intended to be a program that calculates your daily word count against the total word count to see how far along you are in the process.